+++ bricks/basic/src/bootstrap.ts
@@ -9,7 +9,6 @@ import "./dropdown-button/index.js";
 import "./popover/index.js";
 import "./menu/index.js";
 import "./menu-item/index.js";
-import "./nav-menu/index.js";
 import "./tag/index.js";
 import "./tag-list/index.js";
 import "./launchpad/index.js";
