--- bricks/basic/src/nav-menu/site-map/ItemTag.shadow.css
@@ -1,94 +1,0 @@
-.tag-container::part(link) {
-  display: flex;
-  user-select: none;
-  -webkit-user-drag: none;
-  -moz-user-select: none;
-  justify-content: space-between;
-  background-color: var(--color-fill-bg-base-1);
-  border-radius: 2px;
-  padding: 1px 6px;
-}
-
-.tag-container {
-  .tag-text {
-    color: var(--color-normal-text);
-    font-weight: 400;
-  }
-
-  &:hover {
-    .tag-text {
-      color: var(--color-brand-hover);
-    }
-  }
-
-  .tag-suffix {
-    display: inline-flex;
-    font-size: 12px;
-    gap: 4px;
-    align-items: center;
-    color: var(--color-secondary-text);
-  }
-
-  &.is-drag {
-    opacity: 0.4;
-  }
-}
-
-.indicate-wrapper {
-  border: 1px dashed var(--color-brand-hover);
-  border-radius: 4px;
-  height: 20px;
-  margin: 2px 0;
-}
-
-.close {
-  display: inline-flex;
-  margin-top: 2px;
-  padding: 0 2px;
-  &:hover {
-    color: var(--color-error);
-  }
-}
-
-.drag-wrapper {
-  cursor: move;
-  &:hover {
-    color: var(--color-brand-hover);
-  }
-}
-
-.star-icon {
-  display: flex;
-  alignitems: center;
-  color: var(--color-secondary-text);
-  &:hover {
-    color: var(--color-brand-hover);
-  }
-
-  .active {
-    color: var(--palette-yellow-6);
-  }
-}
-
-.recommend-container {
-  &:hover {
-    .star {
-      display: inline-flex;
-    }
-  }
-  .star {
-    display: none;
-  }
-}
-
-.visit-container {
-  &:hover {
-    .operation {
-      display: inline-flex;
-      align-items: center;
-    }
-  }
-  .operation {
-    display: none;
-  }
-}
