--- bricks/basic/src/nav-menu/site-map/processor.spec.ts
@@ -1,601 +1,0 @@
-import { describe, test, expect, jest } from "@jest/globals";
-import { processGroupItems, flatMenuItems } from "./processor.js";
-import { SidebarMenuGroup } from "@next-shared/general/types";
-
-describe("processor", () => {
-  describe("flatMenuItems", () => {
-    it.each([
-      [
-        {
-          title: "服务",
-          type: "group",
-          groupId: "resource",
-          childLayout: "siteMap",
-          items: [
-            {
-              type: "group",
-              title: "网络",
-              items: [
-                {
-                  activeExcludes: [],
-                  activeIncludes: [],
-                  activeMatchSearch: false,
-                  defaultExpanded: false,
-                  exact: false,
-                  icon: {
-                    category: "second-menu",
-                    icon: "loadbalance-second-menu",
-                    lib: "easyops",
-                  },
-                  instanceId: "6026b7c20c8cf",
-                  sort: 0,
-                  text: "负载均衡策略",
-                  to: "/network-resource-monitor/loadbalance",
-                  type: "default",
-                  children: [],
-                  key: "1.2.0",
-                },
-                {
-                  activeExcludes: [],
-                  activeIncludes: [],
-                  activeMatchSearch: false,
-                  defaultExpanded: false,
-                  exact: false,
-                  icon: {
-                    category: "second-menu",
-                    icon: "alibaba-cloud-elastic-public-ip-second-menu",
-                    lib: "easyops",
-                  },
-                  instanceId: "6026b7c20c933",
-                  sort: 0,
-                  text: "阿里云・弹性公网IP",
-                  to: "/network-resource-monitor/ali-eip",
-                  type: "default",
-                  children: [],
-                  key: "1.2.1",
-                },
-              ],
-              key: "1.2",
-            },
-            {
-              type: "group",
-              title: "网络设备",
-              items: [
-                {
-                  activeExcludes: [],
-                  activeIncludes: [],
-                  activeMatchSearch: false,
-                  defaultExpanded: false,
-                  exact: false,
-                  instanceId: "6026b8b05e892",
-                  sort: 10,
-                  text: "交换机",
-                  to: "/network-device-resource-monitor/switch",
-                  type: "default",
-                  children: [],
-                  key: "1.3.0",
-                },
-                {
-                  activeExcludes: [],
-                  activeIncludes: [],
-                  activeMatchSearch: false,
-                  defaultExpanded: false,
-                  exact: false,
-                  instanceId: "6026b8b05e8f6",
-                  sort: 20,
-                  text: "路由器",
-                  to: "/network-device-resource-monitor/router",
-                  type: "default",
-                  children: [],
-                  key: "1.3.1",
-                },
-              ],
-              key: "1.3",
-            },
-            {
-              type: "group",
-              title: "自定义监控",
-              groupFrom: "monitor",
-              items: [
-                {
-                  activeExcludes: [],
-                  activeIncludes: [],
-                  activeMatchSearch: false,
-                  defaultExpanded: false,
-                  exact: false,
-                  instanceId: "6026b8b05e892",
-                  sort: 10,
-                  text: "主机",
-                  to: "/network-device-resource-monitor/host",
-                  type: "default",
-                  children: [],
-                  key: "1.4.0",
-                },
-                {
-                  activeExcludes: [],
-                  activeIncludes: [],
-                  activeMatchSearch: false,
-                  defaultExpanded: false,
-                  exact: false,
-                  instanceId: "6026b8b05e8f6",
-                  sort: 20,
-                  text: "报文",
-                  to: "/network-device-resource-monitor/message",
-                  type: "default",
-                  children: [],
-                  key: "1.4.1",
-                },
-                {
-                  activeExcludes: [],
-                  activeIncludes: [],
-                  activeMatchSearch: false,
-                  defaultExpanded: false,
-                  exact: false,
-                  instanceId: "6026b8b05e891",
-                  sort: 30,
-                  text: "EasyCore",
-                  to: "/network-device-resource-monitor/EasyCore",
-                  type: "default",
-                  children: [],
-                  key: "1.4.2",
-                },
-                {
-                  activeExcludes: [],
-                  activeIncludes: [],
-                  activeMatchSearch: false,
-                  defaultExpanded: false,
-                  exact: false,
-                  instanceId: "6026b8b05e8f7",
-                  sort: 40,
-                  text: "视图",
-                  to: "/network-device-resource-monitor/view",
-                  type: "default",
-                  children: [],
-                  key: "1.4.3",
-                },
-              ],
-              key: "1.4",
-            },
-          ] as any as SidebarMenuGroup[],
-        },
-        [
-          {
-            activeExcludes: [],
-            activeIncludes: [],
-            activeMatchSearch: false,
-            children: [],
-            defaultExpanded: false,
-            exact: false,
-            icon: {
-              category: "second-menu",
-              icon: "loadbalance-second-menu",
-              lib: "easyops",
-            },
-            instanceId: "6026b7c20c8cf",
-            key: "1.2.0",
-            sort: 0,
-            text: "负载均衡策略",
-            to: "/network-resource-monitor/loadbalance",
-            type: "default",
-          },
-          {
-            activeExcludes: [],
-            activeIncludes: [],
-            activeMatchSearch: false,
-            children: [],
-            defaultExpanded: false,
-            exact: false,
-            icon: {
-              category: "second-menu",
-              icon: "alibaba-cloud-elastic-public-ip-second-menu",
-              lib: "easyops",
-            },
-            instanceId: "6026b7c20c933",
-            key: "1.2.1",
-            sort: 0,
-            text: "阿里云・弹性公网IP",
-            to: "/network-resource-monitor/ali-eip",
-            type: "default",
-          },
-          {
-            activeExcludes: [],
-            activeIncludes: [],
-            activeMatchSearch: false,
-            children: [],
-            defaultExpanded: false,
-            exact: false,
-            instanceId: "6026b8b05e892",
-            key: "1.3.0",
-            sort: 10,
-            text: "交换机",
-            to: "/network-device-resource-monitor/switch",
-            type: "default",
-          },
-          {
-            activeExcludes: [],
-            activeIncludes: [],
-            activeMatchSearch: false,
-            children: [],
-            defaultExpanded: false,
-            exact: false,
-            instanceId: "6026b8b05e8f6",
-            key: "1.3.1",
-            sort: 20,
-            text: "路由器",
-            to: "/network-device-resource-monitor/router",
-            type: "default",
-          },
-          {
-            activeExcludes: [],
-            activeIncludes: [],
-            activeMatchSearch: false,
-            children: [],
-            defaultExpanded: false,
-            exact: false,
-            instanceId: "6026b8b05e892",
-            key: "1.4.0",
-            sort: 10,
-            text: "主机",
-            to: "/network-device-resource-monitor/host",
-            type: "default",
-          },
-          {
-            activeExcludes: [],
-            activeIncludes: [],
-            activeMatchSearch: false,
-            children: [],
-            defaultExpanded: false,
-            exact: false,
-            instanceId: "6026b8b05e8f6",
-            key: "1.4.1",
-            sort: 20,
-            text: "报文",
-            to: "/network-device-resource-monitor/message",
-            type: "default",
-          },
-          {
-            activeExcludes: [],
-            activeIncludes: [],
-            activeMatchSearch: false,
-            children: [],
-            defaultExpanded: false,
-            exact: false,
-            instanceId: "6026b8b05e891",
-            key: "1.4.2",
-            sort: 30,
-            text: "EasyCore",
-            to: "/network-device-resource-monitor/EasyCore",
-            type: "default",
-          },
-          {
-            activeExcludes: [],
-            activeIncludes: [],
-            activeMatchSearch: false,
-            children: [],
-            defaultExpanded: false,
-            exact: false,
-            instanceId: "6026b8b05e8f7",
-            key: "1.4.3",
-            sort: 40,
-            text: "视图",
-            to: "/network-device-resource-monitor/view",
-            type: "default",
-          },
-        ],
-      ],
-    ])("should work", (item, result) => {
-      expect(flatMenuItems(item as SidebarMenuGroup)).toEqual(result);
-    });
-  });
-
-  describe("processGroupItems", () => {
-    it.each([
-      [
-        [
-          {
-            type: "group",
-            title: "自定义监控",
-            groupFrom: "monitor",
-            items: [
-              {
-                activeExcludes: [],
-                activeIncludes: [],
-                activeMatchSearch: false,
-                defaultExpanded: false,
-                exact: false,
-                instanceId: "6026b8b05e892",
-                sort: 10,
-                text: "主机",
-                to: "/network-device-resource-monitor/host",
-                type: "default",
-                children: [],
-                key: "1.4.0",
-              },
-              {
-                activeExcludes: [],
-                activeIncludes: [],
-                activeMatchSearch: false,
-                defaultExpanded: false,
-                exact: false,
-                instanceId: "6026b8b05e8f6",
-                sort: 20,
-                text: "报文",
-                to: "/network-device-resource-monitor/message",
-                type: "default",
-                children: [],
-                key: "1.4.1",
-              },
-              {
-                activeExcludes: [],
-                activeIncludes: [],
-                activeMatchSearch: false,
-                defaultExpanded: false,
-                exact: false,
-                instanceId: "6026b8b05e891",
-                sort: 30,
-                text: "EasyCore",
-                to: "/network-device-resource-monitor/EasyCore",
-                type: "default",
-                children: [],
-                key: "1.4.2",
-              },
-              {
-                activeExcludes: [],
-                activeIncludes: [],
-                activeMatchSearch: false,
-                defaultExpanded: false,
-                exact: false,
-                instanceId: "6026b8b05e8f7",
-                sort: 40,
-                text: "视图",
-                to: "/network-device-resource-monitor/view",
-                type: "default",
-                children: [],
-                key: "1.4.3",
-              },
-            ],
-            key: "1.4",
-          },
-          {
-            type: "group",
-            title: "网络",
-            items: [
-              {
-                activeExcludes: [],
-                activeIncludes: [],
-                activeMatchSearch: false,
-                defaultExpanded: false,
-                exact: false,
-                icon: {
-                  category: "second-menu",
-                  icon: "loadbalance-second-menu",
-                  lib: "easyops",
-                },
-                instanceId: "6026b7c20c8cf",
-                sort: 0,
-                text: "负载均衡策略",
-                to: "/network-resource-monitor/loadbalance",
-                type: "default",
-                children: [],
-                key: "1.2.0",
-              },
-              {
-                activeExcludes: [],
-                activeIncludes: [],
-                activeMatchSearch: false,
-                defaultExpanded: false,
-                exact: false,
-                icon: {
-                  category: "second-menu",
-                  icon: "alibaba-cloud-elastic-public-ip-second-menu",
-                  lib: "easyops",
-                },
-                instanceId: "6026b7c20c933",
-                sort: 0,
-                text: "阿里云・弹性公网IP",
-                to: "/network-resource-monitor/ali-eip",
-                type: "default",
-                children: [],
-                key: "1.2.1",
-              },
-            ],
-            key: "1.2",
-          },
-          {
-            type: "group",
-            title: "网络设备",
-            items: [
-              {
-                activeExcludes: [],
-                activeIncludes: [],
-                activeMatchSearch: false,
-                defaultExpanded: false,
-                exact: false,
-                instanceId: "6026b8b05e892",
-                sort: 10,
-                text: "交换机",
-                to: "/network-device-resource-monitor/switch",
-                type: "default",
-                children: [],
-                key: "1.3.0",
-              },
-              {
-                activeExcludes: [],
-                activeIncludes: [],
-                activeMatchSearch: false,
-                defaultExpanded: false,
-                exact: false,
-                instanceId: "6026b8b05e8f6",
-                sort: 20,
-                text: "路由器",
-                to: "/network-device-resource-monitor/router",
-                type: "default",
-                children: [],
-                key: "1.3.1",
-              },
-            ],
-            key: "1.3",
-          },
-        ] as any as SidebarMenuGroup[],
-        [
-          {
-            groupFrom: "monitor",
-            groups: [
-              {
-                groupFrom: "monitor",
-                items: [
-                  {
-                    activeExcludes: [],
-                    activeIncludes: [],
-                    activeMatchSearch: false,
-                    children: [],
-                    defaultExpanded: false,
-                    exact: false,
-                    instanceId: "6026b8b05e892",
-                    key: "1.4.0",
-                    sort: 10,
-                    text: "主机",
-                    to: "/network-device-resource-monitor/host",
-                    type: "default",
-                  },
-                  {
-                    activeExcludes: [],
-                    activeIncludes: [],
-                    activeMatchSearch: false,
-                    children: [],
-                    defaultExpanded: false,
-                    exact: false,
-                    instanceId: "6026b8b05e8f6",
-                    key: "1.4.1",
-                    sort: 20,
-                    text: "报文",
-                    to: "/network-device-resource-monitor/message",
-                    type: "default",
-                  },
-                  {
-                    activeExcludes: [],
-                    activeIncludes: [],
-                    activeMatchSearch: false,
-                    children: [],
-                    defaultExpanded: false,
-                    exact: false,
-                    instanceId: "6026b8b05e891",
-                    key: "1.4.2",
-                    sort: 30,
-                    text: "EasyCore",
-                    to: "/network-device-resource-monitor/EasyCore",
-                    type: "default",
-                  },
-                  {
-                    activeExcludes: [],
-                    activeIncludes: [],
-                    activeMatchSearch: false,
-                    children: [],
-                    defaultExpanded: false,
-                    exact: false,
-                    instanceId: "6026b8b05e8f7",
-                    key: "1.4.3",
-                    sort: 40,
-                    text: "视图",
-                    to: "/network-device-resource-monitor/view",
-                    type: "default",
-                  },
-                ],
-                key: "1.4",
-                title: "自定义监控",
-                type: "group",
-              },
-            ],
-          },
-          {
-            groupFrom: "default",
-            groups: [
-              {
-                items: [
-                  {
-                    activeExcludes: [],
-                    activeIncludes: [],
-                    activeMatchSearch: false,
-                    children: [],
-                    defaultExpanded: false,
-                    exact: false,
-                    icon: {
-                      category: "second-menu",
-                      icon: "loadbalance-second-menu",
-                      lib: "easyops",
-                    },
-                    instanceId: "6026b7c20c8cf",
-                    key: "1.2.0",
-                    sort: 0,
-                    text: "负载均衡策略",
-                    to: "/network-resource-monitor/loadbalance",
-                    type: "default",
-                  },
-                  {
-                    activeExcludes: [],
-                    activeIncludes: [],
-                    activeMatchSearch: false,
-                    children: [],
-                    defaultExpanded: false,
-                    exact: false,
-                    icon: {
-                      category: "second-menu",
-                      icon: "alibaba-cloud-elastic-public-ip-second-menu",
-                      lib: "easyops",
-                    },
-                    instanceId: "6026b7c20c933",
-                    key: "1.2.1",
-                    sort: 0,
-                    text: "阿里云・弹性公网IP",
-                    to: "/network-resource-monitor/ali-eip",
-                    type: "default",
-                  },
-                ],
-                key: "1.2",
-                title: "网络",
-                type: "group",
-              },
-              {
-                items: [
-                  {
-                    activeExcludes: [],
-                    activeIncludes: [],
-                    activeMatchSearch: false,
-                    children: [],
-                    defaultExpanded: false,
-                    exact: false,
-                    instanceId: "6026b8b05e892",
-                    key: "1.3.0",
-                    sort: 10,
-                    text: "交换机",
-                    to: "/network-device-resource-monitor/switch",
-                    type: "default",
-                  },
-                  {
-                    activeExcludes: [],
-                    activeIncludes: [],
-                    activeMatchSearch: false,
-                    children: [],
-                    defaultExpanded: false,
-                    exact: false,
-                    instanceId: "6026b8b05e8f6",
-                    key: "1.3.1",
-                    sort: 20,
-                    text: "路由器",
-                    to: "/network-device-resource-monitor/router",
-                    type: "default",
-                  },
-                ],
-                key: "1.3",
-                title: "网络设备",
-                type: "group",
-              },
-            ],
-          },
-        ],
-      ],
-    ])("should work", (groups, result) => {
-      expect(processGroupItems(groups)).toEqual(result);
-    });
-  });
-});
